Combine methods with config flag, add option for optimal
Add changes that allow both methods (bin width based on data ranges, normalized c_vars based on #samples) to be in the same codebase, toggled by passing a flag on execute. Also add option to read in the optimal solution and thus calculate its fitness. Finally, also parallel evaluation in iAMaLGaM. (and moving data to solution instead of fitness for that)